backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
asyncpg==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.2 